feat: Improve widget management and add spinner for long loading times
This commit introduces two new features to improve the user
experience when working with widgets.

Firstly, widgets are now organized into groups, making it easier
to locate and manage related widgets. Secondly, a spinner has been
added to display when a widget takes a long time to load. This
provides feedback to the user.