chore: Add 'plugin.sh' script with parameterized build, install, run, and clean actions
This commit adds the 'plugin.sh' script to the repository, which
provides a set of actions for building, installing, running, and
cleaning the prject. The script can be invoked with the appropriate
action parameter, such as "./plugin.sh build" to compile the plugin.

This new functionality streamlines the development process and
makes it easier to manage the plugin deployment tasks.