feat: Add display of unknown image generation values
This commit adds the ability to display unknown values used in
image generation that are not recognized by the plugin. These
values are displayed in a table labeled "Unknown" and are
primarily intended to give users access to all of the information
embedded within the image. They can also be useful for development
purposes, allowing developers to identify which configuration
values are not yet recognized by the plugin.