Refactor: Simplify Stable Diffusion parameter parsing
In this commit, the 'parse_sd_parameters' function has been
refactored to simplify the the parsing of parameters. Previously,
a callback function was required to perform this task, but now it
can be done more easily. This change makes it much easier to use
the function and improves the overall readability of the plugin
code.