feat: Add option to force visibility of the plugin
This commit adds a the new feature that allows the user to force
the visibility of the plugin and prioritize its display in the
sidebar. This means that even if other plugins are trying to
display information in the sidebar, SDPromptViewer will have
priority and its prompts and generation information will always
be visible.