feat: Add option to set minimum width of sidebar
Users can now configure a minimum width for the sidebar where the
plugin's information is displayed. This is particularly useful
since EoG's default minimum width is often too narrow, causing
widgets to overlap or disappear off the right side of the screen.
By allowing users to set a configurable minimum width, the
plugin's UI can be adjusted to ensure that all information is
fully visible and easily accessible.