Refactor: Simplify user interface update and optimize widget control.
Unnecessary widget control functions were removed from the plugin's
user interface update mechanism, resulting in a more streamlined and
efficient process. The 'utils_widget.h' file has also been updated
to only contain the essential functions needed for the plugin, which
are now fully documented. While these changes may not be apparent
to the end-user, they greatly improve the quality of the codebase
and make it easier for developers to work with.