feat: Enhance UI customization with CSS for future styling feature
This is the first step towards enabling users to customize the
visual appearance of the plugin. The user interface was improved
by allowing better customization with CSS. It includes a new CSS
file named 'coldmind' as a prototype for future styling options.

While the feature to select visual styles is yet to be implemented,
this commit lays the groundwork for its eventual introduction.