chore: Implement 'make.sh' script for compiling, executing and installing the plugin
In this commit, the 'make.sh' script has been implemented to
automate the compilation, execution and installation process of
the EOG plugin. The script functions similarly to a makefile but
invokes meson commands.

The addition of this script streamlines the building and installation
process of the plugin, saving time and effort for both developers
and end-users.