chore: Implement 'make.sh remove' command for plugin uninstallation
This commit adds the feature of being able to uninstall the plugin
via './make.sh remove'. This option was planned in the previous
commit, but had not been implemented until now.

With this addition, users can now easily remove the plugin if they
no longer need it. This is especially useful for those who are
testing out the plugin and want to remove it from their system.