feat: Add support for multiple languages using .po files and 'update-po' script
This commit adds the ability to localize the text displayed in the
plugin using .po files. The 'update-po' script has also been implemented
to keep all translation files up-to-date with any new translatable text
that may be added to the plugin's code.

This feature will greatly improve the accessibility of the plugin
to users who speak different languages, making it a more versatile
and user-friendly tool.